Rental api

+++ app/Http/Controllers/rentalApiController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers;
 
 use App\Rental;
 use App\Http\Requests;
+use Illuminate\Support\Facades\Validator;
 use League\Fractal\Manager;
 use Illuminate\Http\Request;
 use League\Fractal\Pagination\Cursor;
@@ -13,12 +14,7 @@ use Symfony\Component\HttpFoundation\Response as Status;
 class rentalApiController extends Controller
 {
     // TODO: [v1.0.0] #114 Add notifications for the backend.
-    // TODO: [v1.0.0] #115 Add API documentation.
-    // TODO: [v1.0.0] #116 Add API routes.
     // TODO: [v1.1.0] #117 Add UNIT tests.
-
-    // TODO: [BUG] fix try to get property of non-object. When no rentals and index method is called.
-    // TODO: [BUG] fix Call to a member function delete() on null -> delete controller
 
     public function construct()
     {
@@ -42,15 +38,20 @@ class rentalApiController extends Controller
             $rentals = Rental::take(5)->get();
         }
 
-        $prevCursorStr = $request->input('prevCursor', 6);
-        $newCursorStr  = $rentals->last()->id;
+        if (count($rentals) > 0) {
+            $prevCursorStr = $request->input('prevCursor', 6);
+            $newCursorStr  = $rentals->last()->id;
 
-        $cursor   = new Cursor($currentCursorStr, $prevCursorStr, $newCursorStr, $rentals->count());
-        $resource = new Collection($rentals, $this->parser());
-        $resource->setCursor($cursor);
+            $cursor   = new Cursor($currentCursorStr, $prevCursorStr, $newCursorStr, $rentals->count());
+            $resource = new Collection($rentals, $this->parser());
+            $resource->setCursor($cursor);
 
-        $content = $fractal->createData($resource)->toJson();
-        $status  = Status::HTTP_OK;
+            $content = $fractal->createData($resource)->toJson();
+            $status  = Status::HTTP_OK;
+        } elseif(count($rentals) === 0) {
+            $content = ['message' => 'Er zijn geen verhuringen'];
+            $status  = Status::HTTP_OK;
+        }
 
         return response($content, $status)->header('Content-Type', 'application/json');
     }
@@ -64,9 +65,12 @@ class rentalApiController extends Controller
     public function delete($id)
     {
         $rental = Rental::find($id);
-        $rental->delete();
 
-        switch(count($rental->count())) {
+        if (count($rental) > 0) {
+            $rental->delete();
+        }
+
+        switch(count($rental)) {
             case '1':
                 $status  = Status::HTTP_OK;
                 $content = ['message' => 'rental deleted'];
@@ -89,20 +93,39 @@ class rentalApiController extends Controller
      */
     public function insert(Request $request)
     {
-        $rental             = new Rental;
-        $rental->Start_date = $request->get('Start_datum');
-        $rental->End_date   = $request->get('Eind_datum');
-        $rental->Status     = $request->get('Status');
-        $rental->Email      = $request->get('Email');
-        $rental->telephone  = $request->get('telephone');
-        $rental->save();
+        $validator = Validator::make($request->all(), [
+            'Start_datum' => 'required',
+            'Eind_datum'  => 'required',
+            'Status'      => 'required',
+            'Email'       => 'required',
+            'telephone'   => 'required'
+        ]);
 
-        if ($rental->count() === 0) {
-            $status = Status::HTTP_OK;
-            $content = ['message' => 'Rental successfull added.'];
-        } elseif ($rental->count() > 0) {
+
+        if (! $validator->fails())
+        {
+            $rental             = new Rental;
+            $rental->Start_date = $request->get('Start_datum');
+            $rental->End_date   = $request->get('Eind_datum');
+            $rental->Status     = $request->get('Status');
+            $rental->Email      = $request->get('Email');
+            $rental->telephone  = $request->get('telephone');
+            $rental->save();
+
+            if ($rental->count() === 0) {
+                $status = Status::HTTP_OK;
+                $content = ['message' => 'could not perform the action.'];
+            } elseif ($rental->count() > 0) {
+                $status = Status::HTTP_BAD_REQUEST;
+                $content = ['message' => 'Rental successfull added'];
+            }
+        } else {
             $status = Status::HTTP_BAD_REQUEST;
-            $content = ['message' => 'Could not perform the action'];
+            $content = [
+                'message'      => 'Validation errors',
+                'http_status'  => $status,
+                'errors'       => $validator->errors()->all()
+            ];
         }
 
         return response($content, $status)->header('Content-Type', 'application/json');
@@ -117,20 +140,39 @@ class rentalApiController extends Controller
      */
     public function edit(Request $request, $id)
     {
-        $rental             = new Rental;
-        $rental->Start_date = $request->get('Start_datum');
-        $rental->End_date   = $request->get('Eind_datum');
-        $rental->Status     = $request->get('Status');
-        $rental->Email      = $request->get('Email');
-        $rental->telephone  = $request->get('telephone');
-        $rental->save();
+        $validator = Validator::make($request->all(), [
+            'Start_datum' => 'required',
+            'Eind_datum'  => 'required',
+            'Status'      => 'required',
+            'Email'       => 'required',
+            'telephone'   => 'required'
+        ]);
 
-        if ($rental->count() === 0) {
-            $status = Status::HTTP_OK;
-            $content = ['message' => 'Rental successfull updated.'];
-        } elseif ($rental->count() > 0) {
+
+        if (! $validator->fails())
+        {
+            $rental             = new Rental;
+            $rental->Start_date = $request->get('Start_datum');
+            $rental->End_date   = $request->get('Eind_datum');
+            $rental->Status     = $request->get('Status');
+            $rental->Email      = $request->get('Email');
+            $rental->telephone  = $request->get('telephone');
+            $rental->save();
+
+            if ($rental->count() === 0) {
+                $status = Status::HTTP_OK;
+                $content = ['message' => 'could not perform the action.'];
+            } elseif ($rental->count() > 0) {
+                $status = Status::HTTP_BAD_REQUEST;
+                $content = ['message' => 'Rental successfull updated'];
+            }
+        } else {
             $status = Status::HTTP_BAD_REQUEST;
-            $content = ['message' => 'Could not perform the action'];
+            $content = [
+                'message'      => 'Validation errors',
+                'http_status'  => $status,
+                'errors'       => $validator->errors()->all()
+            ];
         }
 
         return response($content, $status)->header('Content-Type', 'application/json');
